improve all & fix bug & add guest

+++ routes/web.php
@@ -18,9 +18,11 @@ Route::get('/', function () {
     return view('index');
 });
 
-Route::get('/logout', function(){
-    Auth::user()->tokens()->delete();
-    Auth::logout();
+Route::get('/logout', function () {
+    if (Auth::user()) {
+        Auth::user()->tokens()->delete();
+        Auth::logout();
+    }
     return redirect('/');
 });
 
@@ -32,7 +34,7 @@ Route::get('login/{provider}/redirect', [\App\Http\Controllers\Auth\LoginControl
 Route::get('login/{provider}/callback', [\App\Http\Controllers\Auth\LoginController::class, 'handleProviderCallback']);
 
 Route::middleware(['auth:sanctum'])->group(function () {
-    Route::get('/{any?}', function(){
+    Route::get('/{any?}', function () {
         return view('index');
     });
 });
